JUST FINISHED SWITCHING FROM WORKING DX12 RENDERING TO SDL2_RENDERING WITH NO MULTI VIEWPORT SUPPORT.

+++ WindowsApplication/Vector2.cpp
@@ -1,6 +1,6 @@
 #include "Vector2.h"
 
-Vector2::Vector2(int x, int y)
+Vector2::Vector2(float x, float y)
 {
 	this->x = x;
 	this->y = y;
