+++ WindowsApplication/Vector2.h
@@ -2,7 +2,7 @@
 class Vector2
 {
 public:
-	Vector2(int x=0, int y=0);
+	Vector2(float x=0, float y=0);
 	~Vector2();
 
 	float x;
